Refactored and created tests for #26

+++ src/main/java/ex26/PaymentCalculator.java
@@ -11,7 +11,7 @@ import java.lang.Math;
 
 public class PaymentCalculator {
 
-    public int calculateMonthsUntilPaidOff(double i, double b, int p) {
+    public int calculateMonthsUntilPaidOff(double i, double b, double p) {
 
         double x = -(1.0/30) * Math.log(1 + b/p * (1 - Math.pow(1 + i, 30))) / Math.log(1 + i);
